Rename views to routes for FastAPI apps
This is part of the change I want to do for https://github.com/apache/airflow/pull/43015 but going to break it down to keep it manageable.
