+++ airflow/api_fastapi/routes/public/connections.py
@@ -24,8 +24,8 @@ from typing_extensions import Annotated
 from airflow.api_fastapi.db.common import get_session, paginated_select
 from airflow.api_fastapi.openapi.exceptions import create_openapi_http_exception_doc
 from airflow.api_fastapi.parameters import QueryLimit, QueryOffset, SortParam
+from airflow.api_fastapi.routes.router import AirflowRouter
 from airflow.api_fastapi.serializers.connections import ConnectionCollectionResponse, ConnectionResponse
-from airflow.api_fastapi.views.router import AirflowRouter
 from airflow.models import Connection
 
 connections_router = AirflowRouter(tags=["Connection"], prefix="/connections")
