+++ airflow/api_fastapi/routes/public/variables.py
@@ -23,8 +23,8 @@ from typing_extensions import Annotated
 
 from airflow.api_fastapi.db.common import get_session
 from airflow.api_fastapi.openapi.exceptions import create_openapi_http_exception_doc
+from airflow.api_fastapi.routes.router import AirflowRouter
 from airflow.api_fastapi.serializers.variables import VariableBody, VariableResponse
-from airflow.api_fastapi.views.router import AirflowRouter
 from airflow.models.variable import Variable
 
 variables_router = AirflowRouter(tags=["Variable"], prefix="/variables")
