+++ airflow/api_fastapi/routes/ui/__init__.py
@@ -14,21 +14,13 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
 from __future__ import annotations
 
-from airflow.api_fastapi.views.public.connections import connections_router
-from airflow.api_fastapi.views.public.dag_run import dag_run_router
-from airflow.api_fastapi.views.public.dags import dags_router
-from airflow.api_fastapi.views.public.monitor import monitor_router
-from airflow.api_fastapi.views.public.variables import variables_router
-from airflow.api_fastapi.views.router import AirflowRouter
+from airflow.api_fastapi.routes.router import AirflowRouter
+from airflow.api_fastapi.routes.ui.assets import assets_router
+from airflow.api_fastapi.routes.ui.dashboard import dashboard_router
 
-public_router = AirflowRouter(prefix="/public")
+ui_router = AirflowRouter(prefix="/ui")
 
-
-public_router.include_router(dags_router)
-public_router.include_router(connections_router)
-public_router.include_router(variables_router)
-public_router.include_router(dag_run_router)
-public_router.include_router(monitor_router)
+ui_router.include_router(assets_router)
+ui_router.include_router(dashboard_router)
