+++ airflow/api_fastapi/routes/ui/assets.py
@@ -23,7 +23,7 @@ from sqlalchemy.orm import Session
 from typing_extensions import Annotated
 
 from airflow.api_fastapi.db.common import get_session
-from airflow.api_fastapi.views.router import AirflowRouter
+from airflow.api_fastapi.routes.router import AirflowRouter
 from airflow.models import DagModel
 from airflow.models.asset import AssetDagRunQueue, AssetEvent, AssetModel, DagScheduleAssetReference
