+++ tests/api_fastapi/routes/public/test_monitor.py
