+++ tests/api_fastapi/routes/public/test_variables.py
