+++ tests/api_fastapi/routes/ui/test_assets.py
