+++ tests/api_fastapi/routes/ui/test_dashboard.py
